- removed commented part of rqg.scss - Created DEFAULT_RQIDPRIORITY and DEFAULT_RQIDLANG constants and used them to set the values in the template.json - camelCase rqidPriority and rqidLang - renamed IRqgItem.ts to IRqid.ts - remvoed comment block from getItemByRqid - added uuid to RqgItemSheetData - Removed various commented out code that was replaced in the sheet re-work - the Get Item Macro on RQG System tab uses fromUuid -
