+++ src/actors/item-specific/shared/locationNode.ts
@@ -2,6 +2,7 @@ import { emptyPrice, IPhysicalItem } from "../../../data-model/item-data/IPhysic
 import { hasOwnProperty, localize, RqgError } from "../../../system/util";
 import { ItemDataSource } from "@league-of-foundry-developers/foundry-vtt-types/src/foundry/common/data/data.mjs/itemData";
 import { ItemTypeEnum } from "../../../data-model/item-data/itemTypes";
+import { DEFAULT_RQIDLANG, DEFAULT_RQIDPRIORITY } from "../../../data-model/item-data/IRqid";
 
 export type LocationNode = IPhysicalItem & {
   // For the grouping of physical items in a tree structure
@@ -13,8 +14,8 @@ export type LocationNode = IPhysicalItem & {
 export function createItemLocationTree(itemDatas: ItemDataSource[]): LocationNode {
   let locationTree: LocationNode = {
     rqid: "",
-    rqidpriority: 0,
-    rqidlang: "",
+    rqidPriority: DEFAULT_RQIDPRIORITY,
+    rqidLang: DEFAULT_RQIDLANG,
     name: "",
     id: "",
     description: "",
@@ -60,8 +61,8 @@ export function createItemLocationTree(itemDatas: ItemDataSource[]): LocationNod
       }
       return {
         rqid: (itemData.data as any).rqid,
-        rqidpriority: (itemData.data as any).rqidpriority,
-        rqidlang: (itemData.data as any).rqidlang,
+        rqidPriority: (itemData.data as any).rqidPriority,
+        rqidLang: (itemData.data as any).rqidLang,
         name: itemData.name,
         id: itemData._id,
         location: location,
@@ -95,8 +96,8 @@ export function createItemLocationTree(itemDatas: ItemDataSource[]): LocationNod
   const virtualNodes: LocationNode[] = [...virtualNodesMap].map(([i, node]) => {
     return {
       rqid: "",
-      rqidpriority: 0,
-      rqidlang: "",
+      rqidPriority: DEFAULT_RQIDPRIORITY,
+      rqidLang: DEFAULT_RQIDLANG,
       name: node.location,
       id: "",
       quantity: 1,
