--- src/data-model/item-data/IRqgItem.ts
@@ -1,5 +1,0 @@
-export interface IRqgItem {
-  rqid: string;
-  rqidpriority: number;
-  rqidlang: string;
-}
