+++ src/data-model/item-data/armorData.ts
@@ -1,4 +1,5 @@
 import { emptyPrice, IPhysicalItem } from "./IPhysicalItem";
+import { DEFAULT_RQIDLANG, DEFAULT_RQIDPRIORITY } from "./IRqid";
 import { ItemTypeEnum } from "./itemTypes";
 
 export const armorTypeTranslationKeys = [
@@ -62,8 +63,8 @@ export interface ArmorDataProperties {
 
 export const emptyArmor: ArmorDataSourceData = {
   rqid: "",
-  rqidpriority: 0,
-  rqidlang: "",
+  rqidPriority: DEFAULT_RQIDPRIORITY,
+  rqidLang: DEFAULT_RQIDLANG,
   description: "",
   gmNotes: "",
   size: 0,
