+++ src/data-model/item-data/cultData.ts
@@ -1,7 +1,7 @@
 import { Resource } from "../shared/resource";
 import { JournalEntryLink } from "../shared/journalentrylink";
 import { ItemTypeEnum } from "./itemTypes";
-import { IRqgItem } from "./IRqgItem";
+import { DEFAULT_RQIDLANG, DEFAULT_RQIDPRIORITY, IRqgItem } from "./IRqid";
 
 export enum CultRankEnum {
   LayMember = "layMember",
@@ -15,8 +15,8 @@ export enum CultRankEnum {
 
 export interface CultDataSourceData extends JournalEntryLink, IRqgItem {
   rqid: string;
-  rqidpriority: number;
-  rqidlang: string;
+  rqidPriority: number;
+  rqidLang: string;
   rank: CultRankEnum; // TODO You can be a Rune Lord and Priest!
   runePoints: Resource;
   tagline: string;
@@ -48,8 +48,8 @@ export interface CultDataProperties {
 
 export const emptyCult: CultDataSourceData = {
   rqid: "",
-  rqidpriority: 0,
-  rqidlang: "",
+  rqidPriority: DEFAULT_RQIDPRIORITY,
+  rqidLang: DEFAULT_RQIDLANG,
   rank: CultRankEnum.LayMember,
   runePoints: { value: 0, max: 0 },
   tagline: "",
