+++ src/data-model/item-data/gearData.ts
@@ -1,4 +1,5 @@
 import { emptyPrice, IPhysicalItem } from "./IPhysicalItem";
+import { DEFAULT_RQIDLANG, DEFAULT_RQIDPRIORITY } from "./IRqid";
 import { ItemTypeEnum } from "./itemTypes";
 
 export interface GearDataSourceData extends IPhysicalItem {}
@@ -18,8 +19,8 @@ export interface GearDataProperties {
 
 export const emptyGear: GearDataSourceData = {
   rqid: "",
-  rqidpriority: 0,
-  rqidlang: "",
+  rqidPriority: DEFAULT_RQIDPRIORITY,
+  rqidLang: DEFAULT_RQIDLANG,
   description: "",
   gmNotes: "",
   quantity: 1,
