+++ src/data-model/item-data/hitLocationData.ts
@@ -1,7 +1,7 @@
 import { emptyResource, Resource } from "../shared/resource";
 import { ItemTypeEnum } from "./itemTypes";
 import { ActorHealthState } from "../actor-data/attributes";
-import { IRqgItem } from "./IRqgItem";
+import { DEFAULT_RQIDLANG, DEFAULT_RQIDPRIORITY, IRqgItem } from "./IRqid";
 
 // TODO differentiate between severed & maimed? slash / crush or impale
 export const hitLocationHealthStatuses = ["healthy", "wounded", "useless", "severed"] as const;
@@ -15,9 +15,6 @@ export enum HitLocationTypesEnum {
 }
 
 export interface HitLocationDataSourceData extends IRqgItem {
-  rqid: string;
-  rqidpriority: number;
-  rqidlang: string;
   dieFrom: number;
   dieTo: number;
   /** Max and value added by ActorSheet.prepareData */
@@ -53,8 +50,8 @@ export interface HitLocationDataProperties {
 
 export const emptyHitLocation: HitLocationDataSourceData = {
   rqid: "",
-  rqidpriority: 0,
-  rqidlang: "",
+  rqidPriority: DEFAULT_RQIDPRIORITY,
+  rqidLang: DEFAULT_RQIDLANG,
   dieFrom: 0,
   dieTo: 0,
   hitPoints: emptyResource,
