+++ src/data-model/item-data/runeMagicData.ts
@@ -1,4 +1,5 @@
 import { JournalEntryLink } from "../shared/journalentrylink";
+import { DEFAULT_RQIDLANG, DEFAULT_RQIDPRIORITY } from "./IRqid";
 import { ItemTypeEnum } from "./itemTypes";
 import { Spell, SpellConcentrationEnum, SpellDurationEnum, SpellRangeEnum } from "./spell";
 
@@ -31,8 +32,8 @@ export interface RuneMagicDataProperties {
 
 export const emptyRuneMagic: RuneMagicDataSourceData = {
   rqid: "",
-  rqidpriority: 0,
-  rqidlang: "",
+  rqidPriority: DEFAULT_RQIDPRIORITY,
+  rqidLang: DEFAULT_RQIDLANG,
   cultId: "",
   runes: [],
   points: 0,
