+++ src/i18n/en.json
@@ -702,9 +702,9 @@
       },
       "RqgItem": {
         "Error": {
-          "MoreThanOneRqidMatchInWorld": "More than one RQG Item in the world (not in Compendia) matched the rqid \"{rqid}\", rqidlang \"{rqidlang}\", and rqidpriority \"{rqidpriority}\". Please review your RQG System settings.  See console log for item ids.",
-          "MoreThanOneRqidMatchInCompendia": "More than one RQG Item in installed Compendia matched the rqid \"{rqid}\", rqidlang \"{rqidlang}\", and rqidpriority \"{rqidpriority}\". Please review your RQG System settings.  See console log for item ids.",
-          "ItemNotFoundByRqid": "Item not found in world or Compendia matching the rqid \"{rqid}\", and rqidlang \"{rqidlang}\". See console log."
+          "MoreThanOneRqidMatchInWorld": "More than one RQG Item in the world (not in Compendia) matched the rqid \"{rqid}\", rqidLang \"{rqidLang}\", and rqidPriority \"{rqidPriority}\". Please review your RQG System settings.  See console log for item ids.",
+          "MoreThanOneRqidMatchInCompendia": "More than one RQG Item in installed Compendia matched the rqid \"{rqid}\", rqidLang \"{rqidLang}\", and rqidPriority \"{rqidPriority}\". Please review your RQG System settings.  See console log for item ids.",
+          "ItemNotFoundByRqid": "Item not found in world or Compendia matching the rqid \"{rqid}\", and rqidLang \"{rqidLang}\". See console log."
         }
       },
       "Rune": {
@@ -907,14 +907,14 @@
       }
     },
     "RQGSystem": {
-      "rqid": "Rune Quest System ID",
+      "rqid": "RuneQuest System ID",
       "QuickRqid": "Create Rune Quest System ID (rqid) based on current item. Item should be localized in English.",
-      "rqidpriority": "System ID Priority",
-      "rqidlang": "System ID Language",
+      "rqidPriority": "System ID Priority",
+      "rqidLang": "This Item Localized to",
       "ItemFoundryId": "Item Foundry ID",
+      "ItemFoundryUuid": "Item Foundry UUID",
       "OwnerFoundryId": "Owner Foundry ID",
-      "GetOwnedItemMacro": "Get This Owned Item",
-      "GetUnownedItemMacro": "Get This Unowned Item",
+      "GetItemMacro": "Get This Item",
       "GetItemLikeThis": "Get Item Like This",
       "CopyInputValue": "Copy this value to the clipboard"
     },
