+++ src/items/hit-location-item/hitLocationSheet.ts
@@ -69,6 +69,7 @@ export class HitLocationSheet extends RqgItemSheet<
       },
       isGM: getGameUser().isGM,
       ownerId: this.document.actor?.id,
+      uuid: this.document.uuid,
       supportedLanguages: CONFIG.supportedLanguages,
     };
   }
