+++ src/items/passion-item/passionSheet.hbs
@@ -27,24 +27,6 @@
       </div>
     </div>
 
-
-
-{{!-- 
-    <fieldset>
-      <legend>{{localize "RQG.Item.Passion.Name"}}</legend>
-      <label>{{localize "RQG.Item.Passion.Passion"}}<br /><select name="data.passion">
-          {{#select passionData.passion}}
-            {{#each sheetSpecific.passionTypes}}
-              <option
-                value="{{localize (concat 'RQG.Item.Passion.PassionEnum.' this)}}"
-              >{{this}}</option>
-            {{/each}}
-          {{/select}}
-        </select></label>
-      <label>{{localize "RQG.Item.Passion.Subject"}}<br />(
-        <input type="text" name="data.subject" value="{{passionData.subject}}" />
-        )</label>
-    </fieldset> --}}
   </header>
 
   <main>
