+++ src/items/passion-item/passionSheet.ts
@@ -64,6 +64,7 @@ export class PassionSheet extends RqgItemSheet<
       },
       isGM: getGameUser().isGM,
       ownerId: this.document.actor?.id,
+      uuid: this.document.uuid,
       supportedLanguages: CONFIG.supportedLanguages,
     };
   }
