+++ src/items/rune-item/runeSheet.hbs
@@ -110,12 +110,6 @@
             id="has-experience-{{data._id}}"
             {{checked runeData.hasExperience}}>
 
-          {{!-- <label for="chance-{{data._id}}">{{localize "RQG.Item.Rune.PercentageChance"}}</label>
-          <input type="number" id="chance-{{data._id}}" name="data.chance" min="0" max="999" value="{{runeData.chance}}">
-
-          <label for="has-experience-{{data._id}}">{{localize "RQG.Item.Rune.HasExperienceQ"}}</label>
-          <input type="checkbox" id="has-experience-{{data._id}}" name="data.hasExperience" {{checked runeData.hasExperience}}> --}}
-
           <label for="opposing-rune-{{data._id}}">{{localize "RQG.Item.Rune.OpposingRune"}}</label>
           <div class="rune-property">
             <img class="rune" src="{{runeImg runeData.opposingRune}}" title="{{runeData.opposingRune}}"/>
