+++ src/items/rune-magic-item/runeMagicSheet.ts
@@ -72,6 +72,7 @@ export class RuneMagicSheet extends RqgItemSheet<
       },
       isGM: getGameUser().isGM,
       ownerId: this.document.actor?.id,
+      uuid: this.document.uuid,
       supportedLanguages: CONFIG.supportedLanguages,
     };
   }
