+++ src/items/skill-item/skillSheet.hbs
@@ -134,12 +134,6 @@
               style="grid-column: 1 / span 2;">
           </div>
 
-          {{!-- {{localize "RQG.Item.Skill.SorceryRunes"}}<br>
-          {{#each skillData.runes}}
-          <span><img class="rune" src="{{runeImg this}}" title="{{this}}"><input name="data.runes" type="text" value="{{this}}"> </span>
-          {{/each}}
-          <input name="data.runes" type="text" list="allRunes" placeholder="Add new rune"><br> --}}
-
         </div>
 
       </div>
