+++ src/items/spirit-magic-item/spiritMagicSheet.ts
@@ -63,6 +63,7 @@ export class SpiritMagicSheet extends RqgItemSheet<
       },
       isGM: getGameUser().isGM,
       ownerId: this.document.actor?.id,
+      uuid: this.document.uuid,
       supportedLanguages: CONFIG.supportedLanguages,
     };
   }
