+++ src/rqg.scss
@@ -949,26 +949,6 @@ body {
   input {
     width: inherit;
   }
-
-  // &.skill,
-  // &.armor,
-  // &.passion,
-  // &.rune {
-  //   div.paren-head {
-  //     display: inline-flex;
-  //     align-items: center;
-  //   }
-  //   fieldset {
-  //     display: flex;
-  //   }
-  //   .paren-head {
-  //     margin-left: 0.5rem;
-  //   }
-  //   label {
-  //     white-space: nowrap;
-  //     margin-right: 0.6rem;
-  //   }
-  // }
 }
 
 .rqg img.rune {
